Upload: improve error messages
We now distinguish between errors returned from the Flask application
and NGINX errors. If we get a Flask error, we show the request data
since that will be an error message that we defined ourselves. If it
is an NGINX error, the data will be some HTML markup, so we don't wanna
show that. Instead, we just show the status message (e.g. 'Request
entity too large').